Changed the qemu vm vga parameter to a string instead of object, because my cloud init images on proxmox 6.2-4 wouldn't start when set as object.
The hanging images showed "vga type=serial0" in the webinterface. Whereas manually created cloud-init VMs
show "vga serial0". So I noticed something was strange.

When I set it to something else and then back it shows "vga serial0" and the VM boots successfully.

This commit solves the problem, but removes the possibilty to
the set the size of the video memory. I don't need it for my server VMs, so I don't care.